+++ pkg/db/ent/transaction/where.go
@@ -162,14 +162,14 @@ func SignaturePlatform(v string) predicate.Transaction {
 }
 
 // CreatetimeUtc applies equality check predicate on the "createtime_utc" field. It's identical to CreatetimeUtcEQ.
-func CreatetimeUtc(v int) predicate.Transaction {
+func CreatetimeUtc(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldCreatetimeUtc), v))
 	})
 }
 
 // UpdatetimeUtc applies equality check predicate on the "updatetime_utc" field. It's identical to UpdatetimeUtcEQ.
-func UpdatetimeUtc(v int) predicate.Transaction {
+func UpdatetimeUtc(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldUpdatetimeUtc), v))
 	})
@@ -1118,21 +1118,21 @@ func SignaturePlatformContainsFold(v string) predicate.Transaction {
 }
 
 // CreatetimeUtcEQ applies the EQ predicate on the "createtime_utc" field.
-func CreatetimeUtcEQ(v int) predicate.Transaction {
+func CreatetimeUtcEQ(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldCreatetimeUtc), v))
 	})
 }
 
 // CreatetimeUtcNEQ applies the NEQ predicate on the "createtime_utc" field.
-func CreatetimeUtcNEQ(v int) predicate.Transaction {
+func CreatetimeUtcNEQ(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldCreatetimeUtc), v))
 	})
 }
 
 // CreatetimeUtcIn applies the In predicate on the "createtime_utc" field.
-func CreatetimeUtcIn(vs ...int) predicate.Transaction {
+func CreatetimeUtcIn(vs ...int64) predicate.Transaction {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -1149,7 +1149,7 @@ func CreatetimeUtcIn(vs ...int) predicate.Transaction {
 }
 
 // CreatetimeUtcNotIn applies the NotIn predicate on the "createtime_utc" field.
-func CreatetimeUtcNotIn(vs ...int) predicate.Transaction {
+func CreatetimeUtcNotIn(vs ...int64) predicate.Transaction {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -1166,49 +1166,49 @@ func CreatetimeUtcNotIn(vs ...int) predicate.Transaction {
 }
 
 // CreatetimeUtcGT applies the GT predicate on the "createtime_utc" field.
-func CreatetimeUtcGT(v int) predicate.Transaction {
+func CreatetimeUtcGT(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldCreatetimeUtc), v))
 	})
 }
 
 // CreatetimeUtcGTE applies the GTE predicate on the "createtime_utc" field.
-func CreatetimeUtcGTE(v int) predicate.Transaction {
+func CreatetimeUtcGTE(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldCreatetimeUtc), v))
 	})
 }
 
 // CreatetimeUtcLT applies the LT predicate on the "createtime_utc" field.
-func CreatetimeUtcLT(v int) predicate.Transaction {
+func CreatetimeUtcLT(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldCreatetimeUtc), v))
 	})
 }
 
 // CreatetimeUtcLTE applies the LTE predicate on the "createtime_utc" field.
-func CreatetimeUtcLTE(v int) predicate.Transaction {
+func CreatetimeUtcLTE(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldCreatetimeUtc), v))
 	})
 }
 
 // UpdatetimeUtcEQ applies the EQ predicate on the "updatetime_utc" field.
-func UpdatetimeUtcEQ(v int) predicate.Transaction {
+func UpdatetimeUtcEQ(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldUpdatetimeUtc), v))
 	})
 }
 
 // UpdatetimeUtcNEQ applies the NEQ predicate on the "updatetime_utc" field.
-func UpdatetimeUtcNEQ(v int) predicate.Transaction {
+func UpdatetimeUtcNEQ(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldUpdatetimeUtc), v))
 	})
 }
 
 // UpdatetimeUtcIn applies the In predicate on the "updatetime_utc" field.
-func UpdatetimeUtcIn(vs ...int) predicate.Transaction {
+func UpdatetimeUtcIn(vs ...int64) predicate.Transaction {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -1225,7 +1225,7 @@ func UpdatetimeUtcIn(vs ...int) predicate.Transaction {
 }
 
 // UpdatetimeUtcNotIn applies the NotIn predicate on the "updatetime_utc" field.
-func UpdatetimeUtcNotIn(vs ...int) predicate.Transaction {
+func UpdatetimeUtcNotIn(vs ...int64) predicate.Transaction {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -1242,28 +1242,28 @@ func UpdatetimeUtcNotIn(vs ...int) predicate.Transaction {
 }
 
 // UpdatetimeUtcGT applies the GT predicate on the "updatetime_utc" field.
-func UpdatetimeUtcGT(v int) predicate.Transaction {
+func UpdatetimeUtcGT(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldUpdatetimeUtc), v))
 	})
 }
 
 // UpdatetimeUtcGTE applies the GTE predicate on the "updatetime_utc" field.
-func UpdatetimeUtcGTE(v int) predicate.Transaction {
+func UpdatetimeUtcGTE(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldUpdatetimeUtc), v))
 	})
 }
 
 // UpdatetimeUtcLT applies the LT predicate on the "updatetime_utc" field.
-func UpdatetimeUtcLT(v int) predicate.Transaction {
+func UpdatetimeUtcLT(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldUpdatetimeUtc), v))
 	})
 }
 
 // UpdatetimeUtcLTE applies the LTE predicate on the "updatetime_utc" field.
-func UpdatetimeUtcLTE(v int) predicate.Transaction {
+func UpdatetimeUtcLTE(v int64) predicate.Transaction {
 	return predicate.Transaction(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldUpdatetimeUtc), v))
 	})
