+++ pkg/db/ent/walletnode.go
@@ -27,9 +27,9 @@ type WalletNode struct {
 	// LocalIP holds the value of the "local_ip" field.
 	LocalIP string `json:"local_ip,omitempty"`
 	// CreatetimeUtc holds the value of the "createtime_utc" field.
-	CreatetimeUtc int `json:"createtime_utc,omitempty"`
+	CreatetimeUtc int64 `json:"createtime_utc,omitempty"`
 	// LastOnlineTimeUtc holds the value of the "last_online_time_utc" field.
-	LastOnlineTimeUtc int `json:"last_online_time_utc,omitempty"`
+	LastOnlineTimeUtc int64 `json:"last_online_time_utc,omitempty"`
 	// Edges holds the relations/edges for other nodes in the graph.
 	// The values are being populated by the WalletNodeQuery when eager-loading is set.
 	Edges                  WalletNodeEdges `json:"edges"`
@@ -125,13 +125,13 @@ func (wn *WalletNode) assignValues(columns []string, values []interface{}) error
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field createtime_utc", values[i])
 			} else if value.Valid {
-				wn.CreatetimeUtc = int(value.Int64)
+				wn.CreatetimeUtc = value.Int64
 			}
 		case walletnode.FieldLastOnlineTimeUtc:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field last_online_time_utc", values[i])
 			} else if value.Valid {
-				wn.LastOnlineTimeUtc = int(value.Int64)
+				wn.LastOnlineTimeUtc = value.Int64
 			}
 		case walletnode.ForeignKeys[0]:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
